様々な変更、docディレクトリを見て下さい

+++ group10/FixedPointer.java
@@ -3,7 +3,7 @@ import robocode.*;
 
 public class FixedPointer extends AntiGrav{
 
-	public FixedPointer(int inx, int iny, int inweight){
+	public FixedPointer(double inx, double iny, double inweight){
 		x = inx;
 		y = iny;
 		weight = inweight;
